Defensive bash scripting

+++ destroy-default-environment.sh
@@ -3,6 +3,9 @@
 cd terraform
 
 sh tf-destroy.sh ${ENVIRONMENT_NAME} aws-s3-backend.tfconfig
+      
+if [[ $? != 0 ]]; then exit 1; fi
+
 rm -f .env
 
 cd -
